Versioned docs fixes
- Move version data out of site-metadata
    - site-metadata is imported by other projects
    - versions data is built from imported files that aren't present in those other projects
- Fix redirect from /docs (no trailing slash)
- Fix broken "View latest docs" link in VersionCTA
- Fix broken Framework selector
