+++ .storybook/useSiteMetadata.js
@@ -1,3 +1,4 @@
 const siteMetadata = require('../site-metadata.js');
+const versionData = require('../src/util/version-data.js');
 
-export default () => siteMetadata;
+export default () => ({ ...siteMetadata, ...versionData });
