+++ gatsby-config.js
@@ -2,21 +2,25 @@ const path = require('path');
 const { global } = require('@storybook/design-system');
 const siteMetadata = require('./site-metadata');
 const getReleaseBranchUrl = require('./src/util/get-release-branch-url');
+const versionData = require('./src/util/version-data');
 
 require('dotenv').config({
   path: `.env.${process.env.NODE_ENV}`,
 });
 
-const { isLatest, versionString } = siteMetadata;
-
 module.exports = {
-  siteMetadata,
+  siteMetadata: {
+    ...siteMetadata,
+    ...versionData,
+  },
   flags: {
     PRESERVE_WEBPACK_CACHE: true,
     FAST_DEV: true,
     QUERY_ON_DEMAND: true,
   },
-  ...(!isLatest ? { assetPrefix: getReleaseBranchUrl(versionString) } : undefined),
+  ...(!versionData.isLatest
+    ? { assetPrefix: getReleaseBranchUrl(versionData.versionString) }
+    : undefined),
   plugins: [
     'gatsby-plugin-react-helmet',
     'gatsby-plugin-typescript',
@@ -138,14 +142,14 @@ module.exports = {
           '/*': [
             'X-XSS-Protection: 1; mode=block',
             'X-Content-Type-Options: nosniff',
-            ...(!isLatest ? ['Access-Control-Allow-Origin: *'] : []),
+            ...(!versionData.isLatest ? ['Access-Control-Allow-Origin: *'] : []),
           ],
           '/versions.json': [
             'Access-Control-Allow-Origin: *',
             'Access-Control-Allow-Headers: Origin, X-Requested-With, Content-Type, Accept',
           ],
-          ...(!isLatest && {
-            [`/docs/${versionString}/*`]: ['X-Robots-Tag: noindex'],
+          ...(!versionData.isLatest && {
+            [`/docs/${versionData.versionString}/*`]: ['X-Robots-Tag: noindex'],
           }),
         },
         // Do not use the default security headers. Use those we have defined above.
@@ -159,7 +163,7 @@ module.exports = {
         component: require.resolve('./src/components/layout/PageLayout'),
       },
     },
-    ...(isLatest
+    ...(versionData.isLatest
       ? [
           {
             resolve: `gatsby-plugin-sitemap`,
